Add ryzenadj timer, ath11k fix, psmouse fix, convert thinkpad-disable-led and tuned-resume to systemd service

+++ scripts/base/base_install.sh
@@ -11,15 +11,24 @@ $environment_header
 $BASE_ENVIRONMENT
 EOF
 
+# Create local bin
+sudo mkdir -p /usr/local/bin
+sudo cp $SPATH/local-bin/* /usr/local/bin
+
 # Create udev rules
 echo "Applying udev rules..."
 sudo mkdir -p /etc/udev/rules.d
-sudo cp $SPATH/udev-rules/*.rules /etc/udev/rules.d/
+for i in $SPATH/udev-rules/*.rules; do
+	[ ! -e "$i" ] && continue
+	echo "Install $(basename "$i")"
+	sudo cp $i /etc/udev/rules.d/
+done
 for i in $SPATH/udev-rules/*.sh; do
+	[ ! -e "$i" ] && continue
 	filename=$(basename "$i")
 
 	echo "Execute $filename" | indent
-	output=$($SPATH/udev-rules/$filename)
+	output="$($SPATH/udev-rules/$filename)"
 	if [ "$?" = 0 ] && [ ! -z "$output" ]; then
 		sudo tee "/etc/udev/rules.d/${filename%.*}.rules" <<< "$output" | indent-2
 	fi
@@ -30,18 +39,35 @@ echo "Applying module-load..."
 sudo mkdir -p /etc/modules-load.d
 sudo cp $SPATH/modules-load/* /etc/modules-load.d/
 
-# Disable leds for thinkpad, if thinkpad_acpi loaded
-if lsmod | grep thinkpad_acpi &> /dev/null; then
-	echo "Applying thinkpad disable idle led services..."
-	sudo mkdir -p /etc/systemd/system
-	sudo cp $SPATH/systemd/thinkpad-disable-led.service /etc/systemd/system/
-	sudo systemctl enable thinkpad-disable-led |& indent-2
-fi
+# Apply /etc/systemd/system
+echo "Applying systemd services..."
+execute_hook_for() {
+	filename=$(basename "$1")
+	filename="${filename%.sh}"
+	filename="${filename%.service}"
+	filename="${filename%.timer}"
+	filename="${filename}.hook.sh"
 
-if lscpu | grep "AMD Ryzen" > /dev/null; then
-	echo "Installing amd ryzenadj service"
-	sudo cp $SPATH/systemd/ryzenadj-* /etc/systemd/system/
-fi
+	[ ! -e "$SPATH/systemd/$filename" ] && return
+	echo "Execute hook $filename" | indent
+	sudo "$SPATH/systemd/$filename" | indent-2
+}
+for i in $SPATH/systemd/*.service $SPATH/systemd/*.timer; do
+	[ ! -e "$i" ] && continue
+	echo "Install $(basename "$i")" | indent
+	sudo cp "$i" /etc/systemd/system/
+	execute_hook_for "$i"
+done
+for i in $SPATH/systemd/*.service.sh $SPATH/systemd/*.timer.sh; do
+	filename=$(basename "$i")
+
+	echo "Execute $filename" | indent
+	output="$($SPATH/systemd/$filename)"
+	if [ "$?" = 0 ] && [ ! -z "$output" ]; then
+		sudo tee "/etc/systemd/system/${filename%.sh}" <<< "$output" | indent-2
+		execute_hook_for "$i"
+	fi
+done
 
 # Setup locale
 echo "Applying locales..."
@@ -55,11 +81,6 @@ sudo localectl set-locale LANG="$BASE_LOCALE_LANG" LC_CTYPE="$BASE_LOCALE_LC_CTY
 # Enable ssd trim timer
 echo "Enabling ssd trim timer"
 sudo systemctl enable fstrim.timer |& indent
-
-# Inject sleep workarounds
-echo "Applying sleep workarounds..."
-sudo mkdir -p /lib/systemd/system-sleep
-sudo cp $SPATH/systemd-sleep/* /lib/systemd/system-sleep
 
 # Add /etc/sudoers.d/config
 echo "Applying sudoers..."
